simplify location formatter code (fixes some bugs)
